+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/redact/requests/UnredactRequest.java
@@ -1,41 +1,28 @@
 package cloud.pangeacyber.pangea.redact.requests;
 
 import cloud.pangeacyber.pangea.BaseRequest;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+import lombok.experimental.SuperBuilder;
+import lombok.extern.jackson.Jacksonized;
 
+@Data
+@EqualsAndHashCode(callSuper = true)
+@Jacksonized
+@JsonInclude(Include.NON_NULL)
+@SuperBuilder
 public class UnredactRequest<T> extends BaseRequest {
 
+	/** Data to unredact */
+	@NonNull
 	@JsonProperty("redacted_data")
 	T redactedData;
 
+	/** FPE context used to decrypt and unredact data */
 	@JsonProperty("fpe_context")
 	String fpeContext;
-
-	protected UnredactRequest(Builder<T> builder) {
-		this.redactedData = builder.redactedData;
-		this.fpeContext = builder.fpeContext;
-	}
-
-	public T getRedactedData() {
-		return redactedData;
-	}
-
-	public String getFpeContext() {
-		return fpeContext;
-	}
-
-	public static class Builder<T> {
-
-		T redactedData;
-		String fpeContext;
-
-		public Builder(T redactedData, String fpeContext) {
-			this.redactedData = redactedData;
-			this.fpeContext = fpeContext;
-		}
-
-		public UnredactRequest<T> build() {
-			return new UnredactRequest<T>(this);
-		}
-	}
 }